update for this pr Yandex Cloud
@yoavrotems Hello!
Thank you very much for your response!
According to your answer, I really only left the changes in the config.yaml file and deleted everything unnecessary.
I am attaching the results of how I was tested with the current config and what the results were after adding new paths.

Before:
== Summary node ==
18 checks PASS
2 checks FAIL
3 checks WARN
0 checks INFO

After:
== Summary node ==
20 checks PASS
1 checks FAIL
2 checks WARN
0 checks INFO

According to the results, we have only one test left that we do not pass - this is 4.2.6, but we are already working on ensuring that our nodes pass it.

Here I am enclosing instructions on how to set up a demo cluster in our cloud to check it - https://cloud.yandex.com/en/docs/managed-kubernetes/quickstart

but to speed up the time, I can also drop the credentials from the test cluster in a personal message so that you can check it yourself, if necessary